pyproject.toml
[project]
name = "vlc"
version = "0.1.0"
description = "Valencia air quality and weather data pipeline"
requires-python = ">=3.10"
dependencies = [
    "certifi>=2025.11.12",
    "charset-normalizer>=3.4.4",
    "confluent-kafka[schemaregistry,json]>=2.12.2",
    "idna>=3.11",
    "jsonschema>=4.23.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "urllib3>=2.5.0",
]

[tool.setuptools]
packages = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-docker>=3.1.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=scripts",
    "--cov=producer",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["scripts", "producer"]
omit = ["tests/*", "*/test_*.py"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = []

[tool.isort]
profile = "black"
line_length = 120

[dependency-groups]
dev = [
    "jsonschema>=4.25.1",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pyyaml>=6.0.3",
    "ruff>=0.14.6",
]
